tree: move `misc` mod contents to `cql_types` mod
These entities are all related to CQL protocol types, so it makes sense
to put them there.

This annihilates the `misc` module, which is good because the name was
not very descriptive.
